requirements.txt
Flask==2.0.3
gunicorn==20.1.0
Flask-restful==0.3.9
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.5.1
flask_jwt_extended==4.3.1
email-validator==1.1.3
python-dotenv==0.20.0
requests-2.27.1